feat: optimize encryption functions for Caesar and Rot
